requirements.txt
anthropic>=0.25.0
chromadb>=0.4.22
sentence-transformers>=2.7.0
beautifulsoup4>=4.12.0
requests>=2.31.0
youtube-transcript-api>=0.6.2
yt-dlp>=2024.1.0
python-frontmatter>=1.1.0
markdown>=3.6
rich>=13.7.0
typer>=0.12.0
pydantic>=2.7.0
python-dotenv>=1.0.1
duckduckgo-search>=6.1.0
tavily-python>=0.3.3
httpx>=0.27.0
numpy>=1.26.0
tqdm>=4.66.0
